Change blocks affected by hail a bit

+++ common/src/main/java/dev/lukebemish/tempest/impl/data/world/WeatherChunkData.java
@@ -397,7 +397,7 @@ public class WeatherChunkData {
 
     private static boolean tryHailBreak(ServerLevel level, BlockPos toFreeze) {
         var hailEffectState = level.getBlockState(toFreeze);
-        if (hailEffectState.is(Constants.BREAKS_WITH_HAIL)) {
+        if (hailEffectState.is(Constants.BREAKS_WITH_HAIL) && !hailEffectState.is(Constants.SAFE_WITH_HAIL)) {
             level.destroyBlock(toFreeze, true);
             return false;
         }
